+++ EL_FitsData/src/lib/StandardKeyword.cpp
@@ -17,13 +17,12 @@
  *
  */
 
-#include "EL_CfitsioWrapper/StandardKeyword.h"
+#include "EL_FitsData/StandardKeyword.h"
 
 #include <algorithm> // copy_if, find_if
 
 namespace Euclid {
-namespace Cfitsio {
-namespace Header {
+namespace FitsIO {
 
 const std::vector<std::string> StandardKeyword::m_mandatories = { "SIMPLE",   "BITPIX", "NAXIS",  "NAXISn", "END",
                                                                   "XTENSION", "PCOUNT", "GCOUNT", "EXTEND" };
@@ -38,14 +37,14 @@ const std::vector<std::string> StandardKeyword::m_reserveds = {
 
 const std::vector<std::string> StandardKeyword::m_comments = { "COMMENT", "HISTORY" };
 
-const std::map<StandardKeyword::Category, const std::vector<std::string> &> StandardKeyword::byCategory() {
-  return { { StandardKeyword::Category::Mandatory, m_mandatories },
-           { StandardKeyword::Category::Reserved, m_reserveds },
-           { StandardKeyword::Category::Comment, m_comments } };
+const std::map<KeywordCategory, const std::vector<std::string> &> StandardKeyword::byCategory() {
+  return { { KeywordCategory::Mandatory, m_mandatories },
+           { KeywordCategory::Reserved, m_reserveds },
+           { KeywordCategory::Comment, m_comments } };
 }
 
 std::vector<std::string>
-StandardKeyword::filterCategories(const std::vector<std::string> &keywords, Category categories) {
+StandardKeyword::filterCategories(const std::vector<std::string> &keywords, KeywordCategory categories) {
   std::vector<std::string> res;
   const auto it = std::copy_if(keywords.begin(), keywords.end(), res.begin(), [&](const std::string &k) {
     return belongsCategories(k, categories);
@@ -54,7 +53,7 @@ StandardKeyword::filterCategories(const std::vector<std::string> &keywords, Cate
   return res;
 }
 
-bool StandardKeyword::belongsCategories(const std::string &keyword, Category categories) {
+bool StandardKeyword::belongsCategories(const std::string &keyword, KeywordCategory categories) {
   const auto standards = byCategory();
   for (const auto &s : standards) {
     if (categories & s.first) {
@@ -64,7 +63,7 @@ bool StandardKeyword::belongsCategories(const std::string &keyword, Category cat
     }
   }
   // At that point, we know the keyword is not in the selected standard categories.
-  if (categories & Category::User) {
+  if (categories & KeywordCategory::User) {
     for (const auto &s : standards) {
       if (matchesOneOf(keyword, s.second)) { // TODO could smarter and not redo matching
         return false;
@@ -111,6 +110,5 @@ bool StandardKeyword::matchesOneOf(const std::string &keyword, const std::vector
   return pos != refs.end();
 }
 
-} // namespace Header
-} // namespace Cfitsio
+} // namespace FitsIO
 } // namespace Euclid